+++ CHANGELOG.md
@@ -1,5 +1,11 @@
 # Change Log
 
+## Version 0.0.1 (2019-09-05)
+### Added
+- Added exception handling and logging around getting SSM parameters.
+- Added `contains` method to `ParamStore`.
+- Added `store` property to `ParamStoreResult`.
+
 ## Version 0.0.1 (2019-08-27)
 ### Added
 - Initial release.